+++ crates/backend-uzu/src/encodable_block/transformer.rs
@@ -32,8 +32,8 @@ pub struct TransformerEncodeOutput<B: Backend> {
 }
 
 impl<B: Backend> TransformerState<B> {
-    pub fn context_length(&self) -> usize {
-        self.context_length as usize
+    pub fn context_length(&self) -> u32 {
+        self.context_length
     }
 
     pub fn prepare(
